add futured for read delite message

+++ TMessagesProj/src/main/java/org/telegram/messenger/PushListenerController.java
@@ -367,15 +367,16 @@ public class PushListenerController {
                         } else if ("MESSAGE_DELETED".equals(loc_key)) {
                             String messages = custom.getString("messages");
                             String[] messagesArgs = messages.split(",");
-                            LongSparseArray<ArrayList<Integer>> deletedMessages = new LongSparseArray<>();
                             ArrayList<Integer> ids = new ArrayList<>();
                             for (int a = 0; a < messagesArgs.length; a++) {
                                 ids.add(Utilities.parseInt(messagesArgs[a]));
                             }
-                            deletedMessages.put(-channel_id, ids);
-                            NotificationsController.getInstance(currentAccount).removeDeletedMessagesFromNotifications(deletedMessages, false);
-
-                            MessagesController.getInstance(currentAccount).deleteMessagesByPush(dialogId, ids, channel_id);
+                            if (!ZapretConfig.isReadDeletedMessagesEnabled()) {
+                                LongSparseArray<ArrayList<Integer>> deletedMessages = new LongSparseArray<>();
+                                deletedMessages.put(-channel_id, ids);
+                                NotificationsController.getInstance(currentAccount).removeDeletedMessagesFromNotifications(deletedMessages, false);
+                                MessagesController.getInstance(currentAccount).deleteMessagesByPush(dialogId, ids, channel_id);
+                            }
                             if (BuildVars.LOGS_ENABLED) {
                                 FileLog.d(tag + " received " + loc_key + " for dialogId = " + dialogId + " mids = " + TextUtils.join(",", ids));
                             }
